CA-280242: Open destination VDI O_DIRECT (as well as source).
Without O_DIRECT we weren't getting IO errors propagated from the
sendfile(2) call (not even on close). This was causing us not to notice
when copying VDIs to full thinly-provisioned SRs.

Note that the O_DIRECT open is part of the Vhd_format library and
is controlled by the 'use_unbuffered' global variable of
Vhd_format_lwt.File. See xapi-project/ocaml-vhd for the details.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>